hotfixes & suggestions
