+++ src/storage/singletons/Numbers/index.ts
