+++ src/utils/types/numbers/index.ts
@@ -56,7 +56,10 @@ export type ReservedNumbers = {
   stateUpdatedOn: string;
 };
 
-export type DocumentsType = Array<{ name: string; allowedFormats: Array<string> }>;
+export type DocumentsType = Array<{
+  name: string;
+  allowedFormats: Array<string>;
+}>;
 
 export type PortingRequirements = Array<{
   id: string;
@@ -72,3 +75,41 @@ export type PortingRequirements = Array<{
   numbering: { startsWith: string; minDigits: number; maxDigits: number };
   documents: DocumentsType;
 }>;
+
+export type RequestType = {
+  id: number;
+  portId: string;
+  createdOn: string;
+  dueDate: string;
+  kind: string;
+  donor: {
+    id: number;
+    name: string;
+  };
+  status: string;
+  numbers: Array<string>;
+  ranges: [
+    {
+      from: string;
+      to: string;
+    },
+  ];
+};
+
+export type PortingRequestPayload = {
+  numbers?: Array<string>;
+  ranges?: Array<{
+    from: string;
+    to: string;
+  }>;
+  donorId: number;
+  contactEmail: string;
+  dueDate: string;
+  lastName?: string;
+  companyName?: string;
+  houseNumber?: string;
+  houseNumberExt?: string;
+  zipCode?: string;
+  customerId?: string;
+  note?: string;
+};
